# openapi.yml
openapi: 3.1.0
info:
  title: Neue Konzepte BaaS Backend API
  version: 0.1.0
  description: |
    Backend API for the Neue Konzepte BaaS project.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/plots/nearest:
    get:
      tags: [plots]
      summary: Find the nearest plots
      description: |
        Searches across all farmers' plots and returns the nearest ones to a
        search point, nearest first. Public endpoint — no authentication
        required.

        Only plots that are currently available are returned: a plot with a
        rental covering the present moment is omitted, and reappears once that
        rental's period has elapsed.

        Each crop a plot offers carries the season a rental of it there must
        fall within, if any (see `PlotCropOffering`), so a customer can check
        before paying via `POST /api/payments/checkout-sessions` rather than
        finding out from its 409 response.

        The search point is given either as coordinates (`lat` and `lon`
        together) or as a German location (`postalCode` or `city`). Exactly
        one of these two approaches must be used; `lat`/`lon` take
        precedence if both are somehow provided. `postalCode` is matched
        exactly; `city` is matched as a case-insensitive substring against
        the location dataset and only the first match (alphabetically) is
        used to resolve coordinates.
      operationId: findNearestPlots
      security: []
      parameters:
        - name: lat
          in: query
          description: Latitude of the search point. Must be provided together with `lon`.
          schema:
            type: number
            format: double
          example: 48.7758
        - name: lon
          in: query
          description: Longitude of the search point. Must be provided together with `lat`.
          schema:
            type: number
            format: double
          example: 9.1829
        - name: postalCode
          in: query
          description: A German postal code, matched exactly against the location dataset.
          schema:
            type: string
          example: '70173'
        - name: city
          in: query
          description: A German city or locality name, matched as a case-insensitive substring.
          schema:
            type: string
          example: Stuttgart
        - name: limit
          in: query
          description: Maximum number of plots to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          example: 10
        - name: farm
          in: query
          description: |
            Optional farm id. Only that farm's free plots are returned (still
            nearest first), e.g. for a farm's detail page — otherwise a farm
            whose plots aren't among the overall nearest `limit` would look empty.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The nearest plots, ordered by ascending distance.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NearbyPlot'
        '400':
          description: Malformed input, or neither coordinates nor a postal code/city were given.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingLocation:
                  value:
                    error: provide lat and lon, or postalCode or city
                invalidCoordinates:
                  value:
                    error: lat and lon must be numbers
                invalidLimit:
                  value:
                    error: limit must be a number between 1 and 100
        '404':
          description: No location matches the given postal code or city.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: no location found for the given postal code or city
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: internal error
...
  /api/payments/checkout-sessions:
    post:
      tags: [payments]
      summary: Start paying for a plot rental
      description: |
        Opens a Stripe Embedded Checkout session for the authenticated
        customer to pay for renting a plot with a crop. Customers only — a
        farmer or admin receives 403.

        This performs the same validation renting a plot always has — crop
        exists, plot exists, the plot offers that crop and both halves of
        its price are set, the start date is 1 to 60 days out — plus a
        fast-fail availability check, so a customer who could not book
        anyway never reaches Stripe. No rental is created here: one is only
        created once Stripe confirms the payment, via the
        `checkout.session.completed` webhook. Poll
        `GET /api/payments/checkout-sessions/{sessionId}` (using the
        `session_id` Stripe's `return_url` redirects back with) to find out
        when that happened.

        If the crop has a season rule (see `PUT /api/crops/{cropID}/season`),
        the rental's whole period must fall within that season for the
        plot's farm — its own rule if it has one, the default rule otherwise
        — or the request fails with 409. The rental does not need to span
        the whole season, only fit inside it. This is checked again when the
        webhook actually creates the rental; if the season rule changed
        between checkout and payment and the period no longer fits, the
        charge is refunded automatically.
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutSessionRequest'
      responses:
        '201':
          description: The checkout session was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSession'
        '400':
          description: Malformed body, plot id, or crop id, or the start date/message fail validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                invalidPlotID:
                  value:
                    error: invalid plot id
                invalidCropID:
                  value:
                    error: invalid crop id
                invalidStartOrMessage:
                  value:
                    error: start date must be 1 to 60 days from now and message must not be blank
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: not authenticated
        '403':
          description: The authenticated account is not a customer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '404':
          description: No plot or crop with the given id exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: plot or crop not found
        '409':
          description: |
            The crop is not offered by the plot (including if it is offered
            but not fully priced yet), the rental period is outside the
            crop's season, or the plot is already rented for an overlapping
            period.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cropNotOffered:
                  value:
                    error: crop is not offered by this plot
                outsideSeason:
                  value:
                    error: rental period is outside the crop's season
                plotUnavailable:
                  value:
                    error: plot is already rented
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: internal error
  /api/payments/checkout-sessions/{sessionId}:
    get:
      tags: [payments]
      summary: Get a checkout session's status
      description: |
        Returns the status of a checkout session belonging to the
        authenticated customer, identified by its Stripe session id — the
        `session_id` query parameter Stripe's `return_url` redirects back
        with. This is what the frontend polls after the embedded checkout
        iframe reports completion, since the rental itself is only created
        once the `checkout.session.completed` webhook is processed, which
        can lag slightly behind. Customers only — a farmer or admin
        receives 403.
      operationId: getCheckoutSessionStatus
      parameters:
        - name: sessionId
          in: path
          required: true
          description: The Stripe Checkout Session id, not this platform's own id.
          schema:
            type: string
            example: cs_test_a1b2c3
      responses:
        '200':
          description: The checkout session's status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionStatus'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: not authenticated
        '403':
          description: The authenticated account is not a customer, or the session belongs to a different customer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficientPermissions:
                  value:
                    error: insufficient permissions
                notOwner:
                  value:
                    error: checkout session does not belong to you
        '404':
          description: No checkout session exists with that Stripe session id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: checkout session not found
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: internal error
  /api/webhooks/stripe:
    post:
      tags: [payments]
      summary: Stripe webhook delivery
      description: |
        Receives `checkout.session.completed` and `checkout.session.expired`
        events from Stripe. Public endpoint — authenticated by the
        `Stripe-Signature` header against the configured webhook secret
        instead of this platform's own session cookie; Stripe itself is the
        only real caller. Not for use by frontend clients.

        On `checkout.session.completed` this is what actually creates the
        rental request the customer paid for — see
        `POST /api/payments/checkout-sessions`. It is idempotent: Stripe
        retries delivery, so a repeat delivery for an already-processed
        session is a no-op.
      operationId: handleStripeWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A Stripe Event object. Not modeled further here — see the Stripe API reference.
      responses:
        '200':
          description: The event was processed (or intentionally ignored, e.g. an event type this platform does not act on).
        '400':
          description: The Stripe-Signature header failed to verify, or the body could not be read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidSignature:
                  value:
                    error: invalid webhook signature
                unreadableBody:
                  value:
                    error: could not read request body
        '500':
          description: |
            Internal error while processing the event. Stripe will retry
            delivery later, which is the intent for an error that might be
            transient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: internal error
...
components:
...
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly JWT cookie set by `POST /api/auth/login`. Valid for 15 minutes.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
    GeoJSONPolygon:
      type: object
      description: A GeoJSON Polygon geometry.
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          description: An array of linear rings; the first is the exterior boundary.
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
      example:
        type: Polygon
        coordinates:
          - - [9.1798, 48.7758]
            - [9.1808, 48.7758]
            - [9.1808, 48.7768]
            - [9.1798, 48.7768]
            - [9.1798, 48.7758]
...
    NearbyPlot:
      type: object
      required: [id, name, field, farm, coordinates, areaSquareMeters, distanceMeters, crops]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Plot A
        field:
          type: string
          format: uuid
          description: ID of the field this plot belongs to.
        farm:
          type: string
          format: uuid
          description: ID of the farm this plot belongs to; also the id used by `GET /api/farms/{farmID}`.
        coordinates:
          $ref: '#/components/schemas/GeoJSONPolygon'
        areaSquareMeters:
          type: number
          format: double
          description: The plot's area in square meters, computed geodesically from its boundary.
          example: 2340.5
        distanceMeters:
          type: number
          format: double
          description: Distance from the search point to the plot's centroid, in meters.
          example: 1523.42
        crops:
          type: array
          description: |
            The crops this plot actually offers a customer, each with the
            total price for renting this plot with it for its full duration,
            and the season a rental of it here must fall within (if any) so
            a customer can check before paying rather than finding out from
            a 409. A crop the plot lists but that is missing either half of
            its price (the plot's own base rate or the crop's farm-wide
            rate) is left out entirely — it is not a real offer yet.
          items:
            $ref: '#/components/schemas/PlotCropOffering'
    PlotCropOffering:
      type: object
      description: A crop as a customer sees it on a specific plot.
      allOf:
        - $ref: '#/components/schemas/Crop'
        - type: object
          required: [priceCents, season]
          properties:
            priceCents:
              type: integer
              format: int32
              description: |
                The total price in EUR cents for renting this plot with this
                crop for its full duration: `(plot's base rate + the crop's
                farm rate) × plot area × duration in weeks`, computed
                server-side.
              example: 4500
            season:
              $ref: '#/components/schemas/Season'
              nullable: true
              description: |
                The season a rental of this crop on this plot's farm must fall
                within - the farm's own rule if it has one, the default rule
                otherwise. Null if the crop can be rented year-round here.
...
    Crop:
      type: object
      description: |
        A crop in the catalog, with the fixed duration renting it books for.
        Whether it is season-dependent, and which season, is a separate rule -
        see `CropSeasonRule` and `PUT /api/crops/{cropID}/season`.
      required: [id, name, durationMonths]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Tomatoes
        durationMonths:
          type: integer
          format: int32
          description: How long renting this crop books a plot for.
          example: 4
    Season:
      type: object
      description: |
        A recurring yearly calendar window - "Spring: Mar 1 - May 31" - that
        repeats every year regardless of which specific year a rental falls
        in. There is one default set, which an admin maintains and every farm
        starts from; a farmer may also create, edit and delete their own
        seasons at any time, independent of the defaults and of any crop. A
        crop's season rule (see `CropSeasonRule`) points at one of these rows.
      required: [id, name, farmId, startMonth, startDay, endMonth, endDay, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Spring
        farmId:
          type: string
          format: uuid
          nullable: true
          description: Null for one of the global defaults, the farm's id for that farm's own version.
        startMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        startDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        endMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        endDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
          description: |
            A season whose endMonth/endDay is earlier in the year than its
            startMonth/startDay wraps the new year (e.g. Winter: Dec 1 - Feb 28).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateCheckoutSessionRequest:
      type: object
      required: [plotId, cropId, startAt, message]
      properties:
        plotId:
          type: string
          format: uuid
          description: ID of the plot to rent.
          example: ed453eb5-2abb-4fcb-9421-f9b69490fe37
        cropId:
          type: string
          format: uuid
          description: ID of the crop to plant. Must be offered (and priced) by the plot.
          example: 3fa1d9c4-6b2e-4e2a-9a3e-3f2b1c0d4e5f
        startAt:
          type: string
          format: date-time
          description: Must be 1 to 60 days from now.
          example: '2026-10-05T08:00:00Z'
        message:
          type: string
          description: A note to the farmer. Must not be blank.
          example: Looking forward to growing tomatoes this season!
    CheckoutSession:
      type: object
      description: |
        What creating a checkout session hands back to the customer: enough
        to render the embedded payment iframe and a summary above it in one
        round trip. No rental exists yet — one is only created once Stripe
        confirms the payment (see `GET /api/payments/checkout-sessions/{sessionId}`).
      required: [clientSecret, plotName, cropName, priceCents]
      properties:
        clientSecret:
          type: string
          description: Passed to `@stripe/react-stripe-js`'s `EmbeddedCheckoutProvider` to render the payment iframe.
        plotName:
          type: string
          example: Plot A
        cropName:
          type: string
          example: Tomatoes
        priceCents:
          type: integer
          format: int32
          description: The total price in EUR cents that will be charged.
          example: 4500
    CheckoutSessionStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, completed, failed, expired, refunded]
          description: |
            `pending`: awaiting payment. `completed`: paid, `rental` is set.
            `failed`: paid but no rental could be created (e.g. the plot was
            taken by a concurrent payment) — refunded automatically.
            `expired`: the session expired unpaid. `refunded`: paid and a
            rental was created, but the farmer later declined it.
        rental:
          allOf:
            - $ref: '#/components/schemas/Rental'
          description: Set once status is completed or refunded.
    Rental:
      type: object
      required: [id, plotId, cropId, startAt, endAt]
      properties:
        id:
          type: string
          format: uuid
        plotId:
          type: string
          format: uuid
          description: ID of the rented plot.
        cropId:
          type: string
          format: uuid
          description: ID of the crop being planted.
        startAt:
          type: string
          format: date-time
          description: |
            When the rental begins, set from the database clock at booking
            time.
          example: '2026-09-13T14:49:56+02:00'
        endAt:
          type: string
          format: date-time
          description: |
            When the rental ends, exclusive — the crop's duration after
            `startAt`. A rental beginning exactly at another's `endAt` does
            not conflict.
          example: '2027-01-13T14:49:56+01:00'
...
security:
  - accessTokenCookie: []